+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog][keep-cl] and this project adheres
 to [Semantic Versioning][semver] .
 
 ## [Unreleased]
+### Changed
+- Fixed support of [Speedbar][speedbar] to observe Zephir files
 
 ## [0.3.0] - 2017-08-03
 ### Added
@@ -17,7 +19,7 @@ to [Semantic Versioning][semver] .
 ## [0.2.0] - 2017-07-31
 ### Added
 - Added `zephir-mode-open-github` command to open the Zephir Mode GitHub page
-- Added `zephir-open-website-home` command to open the Zephir website
+- Added `zephir-open-website-home` command to open the Zephir web site
 
 ### Changed
 - Removed unused `heredoc` functionality
@@ -31,3 +33,4 @@ to [Semantic Versioning][semver] .
 [0.2.0]: https://github.com/sergeyklay/zephir-mode/compare/0.1.0...0.2.0
 [keep-cl]: http://keepachangelog.com
 [semver]: http://semver.org
+[speedbar]: https://www.gnu.org/software/emacs/manual/html_node/emacs/Speedbar.html